Use nitro instance type to ensure console log is always updated
This also ensures that non-zero exit codes from running the network
verifier script are ignored. This is necessary to be able to properly
check for the end of the userdata script. If any command in the userdata
script results in a non-zero exit code, the cloud-init process on the
isntance will be considered failed and the script execution will stop.