Exit 0 when network validation script runs successfully
In the case where the script completes but one or more endpoints are not
reachable, we should exit with a 0 code instead of 1. This will help
ensure that the run of the docker image does not abort early when
endpoints are not reachable.

This change does mean that it will be more difficult to use the script
programatically outside of the existing osd-network-verifier tooling,
but as there is no use-case for that today this should be the most
reliable option.