Allow users to specify instance root volume encryption key (#58)
* Added kmsKeyID option via new createEC2InstanceInput struct

Added a `--kms-key-id <string` CLI flag that allows the user to
specify which KMS key should be used when encrypting the root
volume of ONV-created instances. This option has also been added
as a parameter (`kmsKeyID`) to the `ValidateEgress()` API function.
Not specifying this key simply defaults back to the cloud account
default KMS key.

Implementation-wise, adding this
option as yet-another parameter to `createEC2Instance()` would have
made that function's signature unreadably long, so I instead
refactored the function such that it now takes in a param struct
of type createEC2InstanceInput.

This addresses OSD-11084, where verifier users requested the ability
to specify which KMS key is used for encryption.

* Improved if-statement readability