Add an egressURL field to GenericError
Currently, it is difficult for consuming repos to determine exactly
which egress URLs are blocked without manually doing regex parsing on an
array of failure strings that we return. This commit solves this by
adding an egressURL field with EgressURL() and GetEgressURLFailures()
methods so that consuming repos can more easily obtain this information.

Signed-off-by: Michael Shen <mshen@redhat.com>