revert probe interface and pass platform as map
a workaround of not changing the probe interface and using a
different template for GetExpandedUserData is to pass a
userDataVariable through the probe interface. In the
GetExpandedUserData, based off if a GCP platform exists, the
function chooses userdata-template or startup-script.
also changed gcp verifier to use GetMachineImageID which sets
default machine image.