Integrate new probe interface into GCP framework (#259)
* add gcp_curl_probe

* change startup-script to split printing and curl

service to print userdata begin runs multiple times despite
succeeding. service to run curl fails. printing end token requires
curl to succeed so it does not start running. crash recovery kernel
arming runs everytime vm is created.

* added shell script to run curl and access config

shell script prints starting and ending token and runs curl. shell
script checks if command was successful. also added external IP so
curl has an output

* add systemd instance termination after set delay

used a systemd script and shell script that gets GCP NAME and ZONE
so the instance self deletes. this ensures that if something
happens to the client, the resource is not left on customer
accounts.

* silence kernel completely and fetch egress URLs

Egress URLs are fetched from GitHub as done in the AWS verifier

* remove target after multi-user

target after multi-user unneeded because other outputs are
completely silenced and wont interfere with probe output

* ignore curl error codes that are network errors

systemd service that runs curl should only fail if bash script
returns an error code, this happens if curl error code is
1-4, 27, 41-43, 45 which means curl failed, not a network failure

* update startup-script and gcp probe comments

* combine GCP curl probe into AWS curl probe

added userDataTemplate to the probe interface so
getExpandedUserData can take in different templates for GCP and AWS

* Update pkg/probes/legacy/legacy.go

Co-authored-by: Alex Vulaj <ajvulaj@gmail.com>

* Update pkg/verifier/aws/entry_point.go

Co-authored-by: Alex Vulaj <ajvulaj@gmail.com>

* clean code

* change startup-script to use all userDataVariables

made minor changes to curl probe to make more readible and added
functionality to startup-script to add cacerts and export proxy
environment variables if specified

* revert probe interface and pass platform as map

a workaround of not changing the probe interface and using a
different template for GetExpandedUserData is to pass a
userDataVariable through the probe interface. In the
GetExpandedUserData, based off if a GCP platform exists, the
function chooses userdata-template or startup-script.
also changed gcp verifier to use GetMachineImageID which sets
default machine image.

* remove unused testing function

* remove unused packages and add comment

* use CPUArchitecture.DefaultInstanceType function

* add error

* Update comment for userDataVariable

Co-authored-by: Anthony Byrne <abyrne@redhat.com>

* Change time unit to seconds

Co-authored-by: Anthony Byrne <abyrne@redhat.com>

* import time

* change get_tokens function to camel case

---------

Co-authored-by: Alex Vulaj <ajvulaj@gmail.com>
Co-authored-by: Anthony Byrne <abyrne@redhat.com>