(OSD-25012) Replace PlatformType Strings with Structs
Created	a new package Platform with Platform structs under pkg/cloud.

Platform structs currently only	contain	an array of two possible struct names
per platform type.  Network verifier uses the normalized string name of a
platform throughout its codebase to identify what the correct platform type
is.  For example, when using the AWS platform the code will use String
comparisons to check if the type is "aws."  This code does not change that
logic but simply encompasses it in a structure.

The String() method for the Platform type will return the currently
expected String name for a platform type.  For example, if using the newly
added AWSHCP Platform type, AWSHCP.String() will give "hostedcluster."

A new suite of tests is also added to validate all the functions that are
part of the Platform package.