Add new Platform Struct (#268)
* (OSD-24920) Update verifier docs to include new flags and probe behavior

Updated the docs to include the following new flags:
cpu-arch
egress-list-location
probe

Also udpated the docs to reflect how egress lists are maintained

* (OSD-25012) Replace PlatformType Strings with Structs

Created	a new package Platform with Platform structs under pkg/cloud.

Platform structs currently only	contain	an array of two possible struct names
per platform type.  Network verifier uses the normalized string name of a
platform throughout its codebase to identify what the correct platform type
is.  For example, when using the AWS platform the code will use String
comparisons to check if the type is "aws."  This code does not change that
logic but simply encompasses it in a structure.

The String() method for the Platform type will return the currently
expected String name for a platform type.  For example, if using the newly
added AWSHCP Platform type, AWSHCP.String() will give "hostedcluster."

A new suite of tests is also added to validate all the functions that are
part of the Platform package.

* changes

* renaming platformTypeStruct to platformTypeName

* The order of the preferred names was backwards

* Changed Platform.PlatformByName() to Platform.ByName()

* (OSD-25012)

Updated the package name of Platform to Cloud such that we will use cloud.Platform rather than platform.Platform.

Removed all instances of String comparison for Platform types and instead use type comparisons.

* Forgot to save a file causing test failure

* Minor change to comment showing we are using cloud.Platform and not helpers.Platform

* Update error handling

* Update docs throughout the codebase

* Update integration to work with new platform type

---------

Co-authored-by: Dakota Long <dalong@dalong-thinkpadp1gen3.remote.csb>