OSD-20513: Add support for the `--noproxy` directive in `curl` (#280)
* OSD-20513: Add support for the `--noproxy` directive in `curl`
This adds an additional `Proxy` config to support passing `--noproxy` to `curl` when
probing urls. The values passed will _not_ be proxied through the configured http/https proxy.

I opted to not update the legacy probe based on it being called "legacy" but am open to go down
that route if we still have use cases for it.

I tested this with the CLI locally and ran the integration tests as well.

On the integration tests, I ran into an aws sdk issue that required bumping the dependencies.

* Address feedback

* Quote vars