OSD-25966: Fix issue with `--egress-list-location` CLI flag (#282)
The `--egress-list-location` flag provided by the CLI was being ignored in the
parsing logic that constructs URLS from given input YAML. This primarily ensures
that if the `ValidateEgressInput` has a populated `EgressListYaml` value, we parse it
and use it. I also moved the flow control around a bit for the decision making about where
to source the data from to make it a bit easier to follow.

I wanted to refactor this further and consolidate the behavior between both AWS and GCP,
but opted to tackle that in a follow on so it was easier to review the bug in isolation.